+++ lib/uml_for_serverpod.dart
@@ -1,4 +1,4 @@
 /// UML Generator for serverpod's YAML-based models
 library uml_for_serverpod;
 
-export 'src/uml_generator.dart';
+export 'src/uml/uml_generator.dart';